pyproject.toml
[project]
name = "hdr-conversion"
version = "0.1.0"
description = "tools for converting HDR images between different formats"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colour-science>=0.4.6",
    "imagecodecs>=2025.11.11",
    "pillow>=11.3.0",
    "pillow-heif>=0.22.0",
    "pyexiftool>=0.5.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=0.25.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "pytest>=9.0.2",
    "rich>=14.2.0",
    "ruff>=0.14.10",
    "pymdown-extensions>=10.19.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/hdrconv"]
